+++ src/mjlab/tasks/backflip/config/__init__.py
@@ -1,0 +1,1 @@
+"""Backflip task configurations."""